+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ruleset/utils/KtlintUtils.kt
@@ -1,10 +1,27 @@
-@file:Suppress("HEADER_MISSING_IN_NON_SINGLE_CLASS_FILE")
+@file:Suppress(
+    "HEADER_MISSING_IN_NON_SINGLE_CLASS_FILE",
+    "Deprecation",
+)
 
 package org.cqfn.diktat.ruleset.utils
 
+import org.cqfn.diktat.common.utils.loggerWithKtlintConfig
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.KtLint.ExperimentalParams
 import com.pinterest.ktlint.core.LintError
+import com.pinterest.ktlint.core.RuleSetProvider
+import mu.KotlinLogging
+import org.intellij.lang.annotations.Language
+
+@Suppress("EMPTY_BLOCK_STRUCTURE_ERROR")
+private val log = KotlinLogging.loggerWithKtlintConfig {}
+
+@Suppress("TYPE_ALIAS")
+val defaultCallback: (lintError: LintError, corrected: Boolean) -> Unit = { lintError, _ ->
+    log.warn("Received linting error: $lintError")
+}
+
+typealias LintErrorCallback = (LintError, Boolean) -> Unit
 
 /**
  * Enables ignoring autocorrected errors when in "fix" mode (i.e. when
@@ -27,3 +44,31 @@ fun ExperimentalParams.ignoreCorrectedErrors(): ExperimentalParams =
             cb(error, false)
         }
     })
+
+
+/**
+ * @param ruleSetProviderRef
+ * @param text
+ * @param fileName
+ * @param cb callback to be called on unhandled [LintError]s
+ * @return formatted code
+ */
+@Suppress("LAMBDA_IS_NOT_LAST_PARAMETER")
+fun format(
+    ruleSetProviderRef: () -> RuleSetProvider,
+    @Language("kotlin") text: String,
+    fileName: String,
+    cb: LintErrorCallback = defaultCallback
+): String {
+    val ruleSets = listOf(ruleSetProviderRef().get())
+    return KtLint.format(
+        ExperimentalParams(
+            text = text,
+            ruleSets = ruleSets,
+            fileName = fileName.removeSuffix("_copy"),
+            script = fileName.removeSuffix("_copy").endsWith("kts"),
+            cb = cb,
+            debug = true,
+        ).ignoreCorrectedErrors()
+    )
+}
